exception handling

+++ src/main/java/com/project/eshop/controllers/ProductsController.java
@@ -1,8 +1,9 @@
 package com.project.eshop.controllers;
 
+import com.project.eshop.controllers.interfaces.InventoryRestInterface;
+import com.project.eshop.exceptions.ItemNotFoundException;
 import com.project.eshop.model.Item;
-import com.project.eshop.repository.ItemsRepository;
-import com.sun.jndi.toolkit.url.Uri;
+import com.project.eshop.repository.InventoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,10 +16,10 @@ import java.util.Optional;
 
 @RestController
 @RequestMapping("products")
-public class ProductsController {
+public class ProductsController implements InventoryRestInterface {
 
     @Autowired
-    ItemsRepository repository;
+    InventoryRepository repository;
 
     @GetMapping("/test")
     public String test(){
@@ -26,26 +27,50 @@ public class ProductsController {
     }
 
     @GetMapping("/getItems")
+    @Override
     public List<Item> getProducts(){
-        return repository.findAll();
+        List<Item> items = repository.findAll();
+        if(items.size() > 0){
+            return items;
+        }
+        throw new ItemNotFoundException("The list is empty, please add some items");
     }
 
+
     @PostMapping("/addItem")
-    public void addItem(@RequestBody Item item){
+    @Override
+    public ResponseEntity addItem(@RequestBody Item item) {
         repository.save(item);
+        return ResponseEntity.status(HttpStatus.OK).build();
     }
 
     @GetMapping("/getItem/{id}")
-    public Optional<Item> findItem(@PathVariable int id){
-        return repository.findById(id);
+    @Override
+    public Item findItem(@PathVariable int id){
+        Optional<Item> optional = repository.findById(id);
+
+        if(optional.isPresent()){
+            return optional.get();
+        }
+        throw new ItemNotFoundException("The item "+id+" does not exist");
     }
 
+
+
     @DeleteMapping("/removeItem/{id}")
+    @Override
     public void removeItem(@PathVariable int id){
-        repository.deleteById(id);
+
+        if(repository.existsById(id)) {
+            repository.deleteById(id);
+        }else{
+            throw new ItemNotFoundException("The item "+id+" does not exists");
+        }
+
     }
 
     @PutMapping("/updateItem")
+    @Override
     public ResponseEntity updateItem(@RequestBody Item item){
         if (repository.existsById(item.getId())) {
             repository.save(item);
